% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plot_samples.R
\name{biplot}
\alias{biplot}
\alias{biplot,matrix-method}
\alias{biplot,SummarizedExperiment-method}
\title{Biplot}
\usage{
biplot(object, ...)

\S4method{biplot}{matrix}(
  object,
  x = pca1,
  y = pca2,
  color = subgroup,
  group = subgroup,
  label = NULL,
  feature_label = feature_name,
  ...,
  sampledata,
  featuredata,
  fixed = list(shape = 15, size = 3),
  nloadings = 0
)

\S4method{biplot}{SummarizedExperiment}(
  object,
  x = pca1,
  y = pca2,
  group = subgroup,
  color = subgroup,
  label = NULL,
  feature_label = feature_name,
  ...,
  fixed = list(shape = 15, size = 3),
  nloadings = 0
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{...}{additional svars mapped to aesthetics}

\item{x}{pca1, etc.}

\item{y}{pca2, etc.}

\item{color}{svar mapped to color (symbol)}

\item{label}{svar mapped to label (symbol)}

\item{feature_label}{fvar mapped to (loadings) label}

\item{fixed}{fixed plot aesthetics}

\item{nloadings}{number of loadings per half-axis to plot}
}
\value{
ggplot object
}
\description{
Biplot
}
\examples{
# Read data
    file <- download_data('halama18.metabolon.xlsx')
    object <- read_metabolon(file, plot = FALSE)
# biplot.SummarizedExperiment
    biplot(object)
    biplot(object, x=pca1, y=pca2)
    biplot(object, x=pls1, y=pls2)
    biplot(object, x=lda1, y=lda2)
    biplot(object, x=pca3, y=pca4)
    biplot(object, nloadings = 1)
    biplot(object, color = TIME_POINT)
    biplot(object, color = NULL)
# biplot.matrix
    object \%<>\% exprs()
}
